Job post
When you have clicked yourself into the post it looks well enough. On
the blog, however, it looks like shit. I’ve tried to get the 3 pictures
to be shown as they do in the post but no success. I’ve asked Tikka and
Daniel. Do you have a solution or should I just design it in another
way?